api/requirements.txt
aiohttp==3.7.4.post0
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
bokeh==2.3.2
Bottleneck==1.3.2
branca==0.4.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
dask==2.30.0
dataprep==0.2.15
decorator==5.0.7
defusedxml==0.7.1
Django==3.2.3
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.2
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.1
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
folium==0.12.1
fsspec==2021.4.0
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.7
GitPython==3.1.14
googlemaps==4.4.5
greenlet==1.0.0
idna==2.10
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jason==0.1.7
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
mysql-client==0.0.1
mysql-connector-python==8.0.24
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pgeocode==0.3.0
phonenumbers==8.12.22
pickleshare==0.7.5
Pillow==8.2.0
ply==3.11
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8.1
pydeck==0.6.2
pygeocoder==1.2.5
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
reverse-geocode==1.4.1
scipy==1.6.3
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.13
sqlparse==0.4.1
streamlit==0.81.1
terminado==0.9.4
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
validate-email==1.3
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
xlrd==2.0.1
yarl==1.6.3
